requirements.txt
qdrant-client
ollama
requests
httpx
pydantic
python-Levenshtein
numpy
psycopg2-binary 